disable cosmos
