+++ pom.xml
@@ -69,12 +69,6 @@
 			<artifactId>modelmapper</artifactId>
 			<version>${modelmapper-version}</version>
 		</dependency>
-		<!-- https://mvnrepository.com/artifact/org.json/json -->
-		<dependency>
-			<groupId>org.json</groupId>
-			<artifactId>json</artifactId>
-			<version>${json.version}</version>
-		</dependency>
 		<dependency>
 			<groupId>org.modelmapper.extensions</groupId>
 			<artifactId>modelmapper-jackson</artifactId>
