+++ src/main/java/it/gov/pagopa/spontaneouspayment/config/ConsmosDBConfiguration.java
@@ -9,17 +9,23 @@ import com.azure.spring.data.cosmos.config.AbstractCosmosConfiguration;
 import com.azure.spring.data.cosmos.config.CosmosConfig;
 import com.azure.spring.data.cosmos.core.ResponseDiagnostics;
 import com.azure.spring.data.cosmos.core.ResponseDiagnosticsProcessor;
+import com.azure.spring.data.cosmos.core.mapping.EnableCosmosAuditing;
 import com.azure.spring.data.cosmos.repository.config.EnableCosmosRepositories;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.lang.Nullable;
 
 @Configuration
-@EnableCosmosRepositories
+@EnableCosmosRepositories("it.gov.pagopa.spontaneouspayment.repository")
+@EnableCosmosAuditing
+@EnableConfigurationProperties
+@ConditionalOnExpression("'${info.properties.environment}'!='test'")
 @Slf4j
-public class AppConfiguration extends AbstractCosmosConfiguration {
+public class ConsmosDBConfiguration extends AbstractCosmosConfiguration {
 
 
     @Value("${azure.cosmos.uri}")
