+++ src/main/java/it/gov/pagopa/spontaneouspayment/service/client/ExternalServiceClient.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.spontaneouspayment.service;
+package it.gov.pagopa.spontaneouspayment.service.client;
 
 import java.net.URI;
