+++ src/test/java/it/gov/pagopa/spontaneouspayment/service/ServicesServiceTest.java
@@ -1,9 +1,31 @@
 package it.gov.pagopa.spontaneouspayment.service;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-import static org.mockito.Mockito.spy;
+import com.azure.cosmos.CosmosAsyncClient;
+import com.azure.cosmos.CosmosClientBuilder;
+import com.azure.cosmos.models.CosmosContainerResponse;
+import com.azure.cosmos.models.CosmosDatabaseResponse;
+import it.gov.pagopa.spontaneouspayment.entity.Organization;
+import it.gov.pagopa.spontaneouspayment.entity.Service;
+import it.gov.pagopa.spontaneouspayment.entity.ServiceProperty;
+import it.gov.pagopa.spontaneouspayment.entity.ServiceRef;
+import it.gov.pagopa.spontaneouspayment.exception.AppException;
+import it.gov.pagopa.spontaneouspayment.model.enumeration.PropertyType;
+import it.gov.pagopa.spontaneouspayment.model.enumeration.Status;
+import it.gov.pagopa.spontaneouspayment.repository.OrganizationRepository;
+import it.gov.pagopa.spontaneouspayment.repository.ServiceRepository;
+import org.junit.Rule;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.rules.TemporaryFolder;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpStatus;
+import org.testcontainers.containers.CosmosDBEmulatorContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -15,35 +37,8 @@ import java.security.cert.CertificateException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Rule;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.rules.TemporaryFolder;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
-import org.testcontainers.containers.CosmosDBEmulatorContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
-
-import com.azure.cosmos.CosmosAsyncClient;
-import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.models.CosmosContainerResponse;
-import com.azure.cosmos.models.CosmosDatabaseResponse;
-
-import it.gov.pagopa.spontaneouspayment.entity.Organization;
-import it.gov.pagopa.spontaneouspayment.entity.Service;
-import it.gov.pagopa.spontaneouspayment.entity.ServiceProperty;
-import it.gov.pagopa.spontaneouspayment.entity.ServiceRef;
-import it.gov.pagopa.spontaneouspayment.exception.AppException;
-import it.gov.pagopa.spontaneouspayment.model.enumeration.PropertyType;
-import it.gov.pagopa.spontaneouspayment.model.enumeration.Status;
-import it.gov.pagopa.spontaneouspayment.repository.OrganizationRepository;
-import it.gov.pagopa.spontaneouspayment.repository.ServiceRepository;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.spy;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @SpringBootTest
@@ -53,10 +48,10 @@ class ServicesServiceTest {
 	@Rule
 	public TemporaryFolder tempFolder = new TemporaryFolder();
 
-	@Autowired
+	@MockBean
 	private OrganizationRepository ciRepository;
 
-	@Autowired
+	@MockBean
 	private ServiceRepository serviceRepository;
 
 	@Container
@@ -85,12 +80,12 @@ class ServicesServiceTest {
 		// creation of the database and containers
 		CosmosDatabaseResponse databaseResponse = client.createDatabaseIfNotExists("db").block();
 
-		Assertions.assertEquals(201, databaseResponse.getStatusCode());
+		assertEquals(201, databaseResponse.getStatusCode());
 		CosmosContainerResponse containerResponse = client.getDatabase("db")
 				.createContainerIfNotExists("creditor_institutions", "/fiscalCode").block();
-		Assertions.assertEquals(201, containerResponse.getStatusCode());
+		assertEquals(201, containerResponse.getStatusCode());
 		containerResponse = client.getDatabase("db").createContainerIfNotExists("services", "/fiscalCode").block();
-		Assertions.assertEquals(201, containerResponse.getStatusCode());
+		assertEquals(201, containerResponse.getStatusCode());
 
 		// loading the database with test data
 		Organization ci = new Organization();
