+++ src/test/resources/application.properties
@@ -2,11 +2,11 @@
 info.application.name=@project.artifactId@
 info.application.version=@project.version@
 info.application.description=@project.description@
-info.properties.environment=JUNIT
+info.properties.environment=test
 
 # Cosmos account config
 azure.cosmos.uri=https://pagopa-d-weu-gps-cosmos-account.documents.azure.com:443/
-azure.cosmos.key=f7162UkE9ZixAZZ4ln86dzZOQ2Im68Q6IYah56MpsBTpuydDo3Btd3T5hmlPAgq9jkGIfD43TnUhm6COokcVTA==
+azure.cosmos.key=secret
 azure.cosmos.database=db
 azure.cosmos.populate-query-metrics=false
